Функциональность брони: При создании брони резервируется место на сеанс. При оплате кол-во зарезервированных мест на сеанс уменьшается,а наполненность зала увеличивается. При попытке удалить оплаченную бронь выбрасывается исключение или бронь удаляется и резерв места с сеанса убирается.

+++ src/main/kotlin/ru/evreke/demo/controllers/api/v1/BookingApi.kt
@@ -2,6 +2,7 @@ package ru.evreke.demo.controllers.api.v1
 
 import org.springframework.web.bind.annotation.*
 import ru.evreke.demo.entity.Booking
+import ru.evreke.demo.exceptions.AlreadyPayedException
 import ru.evreke.demo.exceptions.NotFoundException
 import ru.evreke.demo.repository.BookingRepository
 import ru.evreke.demo.repository.MovieSessionRepository
@@ -24,7 +25,9 @@ class BookingApi(
         @RequestParam movieSessionId: Long,
         @RequestParam userId: Long
     ) {
-        val movieSession = movieSessionRepo.findById(movieSessionId).orElseThrow { NotFoundException("Movie session with id=$movieSessionId not found") }
+        val movieSession = movieSessionRepo.findById(movieSessionId).orElseThrow { NotFoundException("Movie session with id=$movieSessionId not found") }.also {
+            it.booked++
+        }
         val user = userRepo.findById(userId).orElseThrow { NotFoundException("User with id=$userId not found") }
         repo.save(Booking().also {
             it.session = movieSession
@@ -32,11 +35,34 @@ class BookingApi(
         })
     }
 
+    @PutMapping("/{id}")
+    fun setPayed(
+        @PathVariable id: Long
+    ) {
+        val booking = repo.findById(id).orElseThrow { NotFoundException("Booking with id=$id not found") }
+        if (booking.payed) {
+            throw AlreadyPayedException("Booking with id=$id was already payed")
+        } else {
+            booking.apply {
+                payed = !payed
+                session!!.booked--
+                session!!.occupancy++
+            }
+            repo.save(booking)
+        }
+    }
+
     @DeleteMapping("/{id}")
     fun deleteBooking(
         @RequestParam(required = false) userId: Long?,
         @PathVariable id: Long
     ) {
-        userId?.let { repo.deleteBookingByUserIdAndId(userId, id) } ?: repo.deleteById(id)
+        val booking = repo.findById(id).orElseThrow { NotFoundException("Booking with id=$id not found") }
+        if (booking.payed) {
+            throw AlreadyPayedException("Booking with id=$id was payed. You shouldn't delete it!")
+        } else {
+            movieSessionRepo.save(booking.session!!.also { it.booked-- })
+            userId?.let { repo.deleteBookingByUserIdAndId(userId, id) } ?: repo.deleteById(id)
+        }
     }
 }
